Use QuarkusMock.install approach instead of global KeycloakCryptomatorVaultsHelperMock.
# Conflicts:
#	backend/src/test/java/org/cryptomator/hub/api/VaultResourceIT.java
